+++ app/random-id/src/components/home-page.tsx
@@ -1,27 +1,48 @@
-import { Box, Button, Container, Flex, Heading, Stack, Text } from '@chakra-ui/react'
-import { nanoid } from 'nanoid'
+import { Box, Button, Checkbox, Container, Flex, Heading, Stack, Text } from '@chakra-ui/react'
 import { NextPage } from 'next'
 import NextHead from 'next/head'
-import { useEffect, useState } from 'react'
+import { FunctionComponent, useEffect, useState } from 'react'
 
+import { generateId, GenerateIdOptions, getDefaultOptions } from 'mvp-app-random-id/src/components/id-generator'
 import { AppFooter } from 'mvp-common-components/src/app-footer'
+
+interface OptionGroupProps {
+  readonly options: GenerateIdOptions
+  readonly setOptions: (nextOptions: GenerateIdOptions) => void
+}
+
+const OptionGroup: FunctionComponent<OptionGroupProps> = ({ options, setOptions }) => {
+  const { hasLetters, hasNumbers } = options
+
+  return (
+    <>
+      <Checkbox isChecked={hasLetters} onChange={() => setOptions({ ...options, hasLetters: !hasLetters })} margin={2}>
+        Letters
+      </Checkbox>
+      <Checkbox isChecked={hasNumbers} onChange={() => setOptions({ ...options, hasNumbers: !hasNumbers })} margin={2}>
+        Numbers
+      </Checkbox>
+    </>
+  )
+}
 
 const useHomePage = () => {
   const [id, setId] = useState<string | undefined>()
+  const [options, setOptions] = useState<GenerateIdOptions>(getDefaultOptions)
 
-  const generateId = () => {
-    setId(nanoid())
+  const setRandomId = () => {
+    setId(generateId(options))
   }
 
   useEffect(() => {
-    generateId()
-  }, [])
+    setId(generateId(options))
+  }, [options])
 
-  return { id, generateId }
+  return { id, options, setOptions, setRandomId }
 }
 
 export const HomePage: NextPage = () => {
-  const { id, generateId } = useHomePage()
+  const { id, options, setOptions, setRandomId } = useHomePage()
 
   return (
     <>
@@ -38,9 +59,10 @@ export const HomePage: NextPage = () => {
             {id ?? <>&nbsp;</>}
           </Box>
           <Flex wrap='wrap' justifyContent='center'>
-            <Button onClick={generateId} colorScheme='primary' margin={1}>
+            <Button onClick={setRandomId} colorScheme='primary' margin={2}>
               Generate
             </Button>
+            <OptionGroup options={options} setOptions={setOptions} />
           </Flex>
         </Stack>
         <AppFooter />