+++ app/random-id/CHANGELOG.md
@@ -5,3 +5,4 @@
 ### Added
 
 - Added home page in which user can generate random id
+- Added options to control the alphabet of id